refactor(scripts): esbuild > webpack
